Refactor logic related to current number
The current number should not have a default value of "0". Instead
its value should be an empty string and only updates once the user
presses digits.
Additionally if the current number is empty then an operator button
should not do anything since it has no number to perform an
operation on.
Added validation to update the operator currently in the expression
only if the expression length is 0 (i.e. contains only a number
and an operator)
Fixed syntax issue in the isValidExpression and evaluateExpression
functions as the numbers were not destructuring as I intended.